requirements.txt
Keras==2.4.3
streamlit==0.85.0
numpy==1.19.5
Pillow==8.2.0
urllib3==1.26.5
scikit-image==0.18.2